requirements.txt
python-dotenv
pykernel
langchain_community
langchain-core
streamlit
numpy
pygame
pillow
panda3d